+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Motions list
 - Dispatch queue list and Scheduled list on Dispatch tab
 
+#### Fixed
+- Collected winning data on auction tab
+
 ## [0.1.0] - 2021-11-26
 #### Added
 - Collection loader